% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TET2.R
\name{TET2}
\alias{TET2}
\title{Fit linear models and/or contrasts to data from Figueroe et al. 2010.}
\usage{
TET2(eset, design, ...)
}
\arguments{
\item{eset}{an ExpressionSet of HELP data (e.g. data(DNAme))}

\item{design}{a formula for a linear model to fit to the data with limma}

\item{...}{any other parameters to pass to lmFit (e.g. method="robust")}
}
\value{
\preformatted{     results from eBayes(lmFit(eset, design, ...))
}
}
\description{
Fit linear models and/or contrasts to data from Figueroe et al. 2010.
}
\examples{

 data(DNAme, package="WorldsSimplestCodeReview")
 covariates <- pData(DNAme) # pData == "pheno-data", a very old function
 library(limma) 

 # let's fit a model with the specified bone marrow blast \% (but beware...)
 design <- with(covariates, model.matrix( ~ IDH:purity + TET2:purity + male))
 fitWithPurityAndSex <- TET2(DNAme, design, method="robust")

 # if you look at DNAme$purity, it doesn't make a lot of sense
 # so let's compare against normal bone marrows (NBM) instead:
 DNAme$NBM <- grepl("NBM", DNAme$title) 
 DNAme$AML <- !DNAme$NBM # the non-NBMs are all AMLs 
 design2 <- with(pData(DNAme), model.matrix( ~ IDH:AML + TET2:AML + AML))
 fitAgainstNBMs <- TET2(DNAme, design2) # AML-specific hyper now in coef 4!
 colnames(design2) # for reference below

 message("IDH vs. non-mutant AML @ p_adj < 0.05:")
 nrow(topTable(fitAgainstNBMs, coef=2, p.val=0.05, n=Inf)))

 message("TET2 vs. non-mutant AML @ p_adj < 0.05:")
 nrow(topTable(fitAgainstNBMs, coef=3, p.val=0.05, n=Inf))

 message("AML vs. normal bone marrow @ p_adj < 0.05:")
 nrow(topTable(fitAgainstNBMs, coef=4, p.val=0.05, n=Inf))

}
